+++ src/main/java/lotto/domain/AutoLottoStrategy.java
@@ -1,7 +1,7 @@
 /*
  * 로또 자동 생성 전략을 나타내는 클래스
  * */
-package lotto;
+package lotto.domain;
 
 import java.util.ArrayList;
 import java.util.Collections;
