+++ src/main/java/lotto/domain/LottoStrategy.java
@@ -1,7 +1,7 @@
 /*
  * 로또 생성 전략을 나타내는 함수형 인터페이스
  * */
-package lotto;
+package lotto.domain;
 
 import java.util.List;
