+++ src/main/java/lotto/util/InputView.java
@@ -1,7 +1,7 @@
 /*
  * UI 입력을 담당하는 클래스
  * */
-package lotto;
+package lotto.util;
 
 import java.io.BufferedReader;
 import java.io.IOException;
