+++ src/main/java/lotto/util/LottoFactory.java
@@ -1,7 +1,12 @@
 /*
  * 로또 한장과, 로또의 묶음단위, 당첨 도메인들의 생성을 담당하는 팩토리 클래스
  * */
-package lotto;
+package lotto.util;
+
+import lotto.domain.Lotto;
+import lotto.domain.LottoStrategy;
+import lotto.domain.Lottos;
+import lotto.domain.Winning;
 
 import java.util.ArrayList;
 import java.util.List;
