+++ src/test/java/lotto/domain/AutoLottoStrategyTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 import org.junit.jupiter.api.Test;
